+++ src/apps/shared/_ConsoleWrapper.ts
@@ -1,5 +1,5 @@
 // this file is a modification of https://github.com/tauri-apps/tauri-plugin-log/blob/v2/guest-js/index.ts
-import { _invoke } from './_tauri';
+import { _invoke, WebviewInformation } from './_tauri';
 
 export interface LogOptions {
   file?: string;
@@ -40,73 +40,11 @@ export enum LogLevel {
   Error,
 }
 
-function getCallerLocation(stack?: string): string | undefined {
-  if (!stack) {
-    return;
-  }
-
-  if (stack.startsWith('Error')) {
-    // Assume it's Chromium V8
-    //
-    // Error
-    //     at baz (filename.js:10:15)
-    //     at bar (filename.js:6:3)
-    //     at foo (filename.js:2:3)
-    //     at filename.js:13:1
-
-    const lines = stack.split('\n');
-    // Find the third line (caller's caller of the current location)
-    const callerLine = lines[3]?.trim();
-    if (!callerLine) {
-      return;
-    }
-
-    const regex =
-      /at\s+(?<functionName>.*?)\s+\((?<fileName>.*?):(?<lineNumber>\d+):(?<columnNumber>\d+)\)/;
-    const match = callerLine.match(regex);
-
-    if (match) {
-      const { functionName, fileName, lineNumber, columnNumber } = match.groups as {
-        functionName: string;
-        fileName: string;
-        lineNumber: string;
-        columnNumber: string;
-      };
-      return `${functionName}@${fileName}:${lineNumber}:${columnNumber}`;
-    } else {
-      // Handle cases where the regex does not match (e.g., last line without function name)
-      const regexNoFunction = /at\s+(?<fileName>.*?):(?<lineNumber>\d+):(?<columnNumber>\d+)/;
-      const matchNoFunction = callerLine.match(regexNoFunction);
-      if (matchNoFunction) {
-        const { fileName, lineNumber, columnNumber } = matchNoFunction.groups as {
-          fileName: string;
-          lineNumber: string;
-          columnNumber: string;
-        };
-        return `<anonymous>@${fileName}:${lineNumber}:${columnNumber}`;
-      }
-    }
-  } else {
-    // Assume it's Webkit JavaScriptCore, example:
-    //
-    // baz@filename.js:10:24
-    // bar@filename.js:6:6
-    // foo@filename.js:2:6
-    // global code@filename.js:13:4
-
-    const traces = stack.split('\n').map((line) => line.split('@'));
-    const filtered = traces.filter(([name, location]) => {
-      return name && name.length > 0 && location !== '[native code]';
-    });
-    // Find the third line (caller's caller of the current location)
-    return filtered[2]?.filter((v) => v.length > 0).join('@');
-  }
-
-  return;
-}
-
+const webviewInfo = new WebviewInformation();
 async function log(level: LogLevel, message: string, options?: LogOptions): Promise<void> {
-  const location = getCallerLocation(new Error().stack);
+  // we use the webview label as the location, intead call stack as the stack on the Seelen UI case
+  // will be always the same because of the console wrapper, this is different from the tauri plugin
+  const location = webviewInfo.label;
 
   const { file, line, keyValues } = options ?? {};